Buildathon_MVP/requirements.txt
#streamlit>=1.24.0
#openai>=1.0.0

streamlit>=1.24.0
google-genai>=0.7.0
google-cloud-speech>=2.10.0

google-generativeai